% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_sample_data.R
\name{dep_sample_data}
\alias{dep_sample_data}
\title{Create Sample Data}
\usage{
dep_sample_data(index)
}
\arguments{
\item{index}{A character scalar or vector listing deprivation measures
to return. These include the area deprivation index (\code{"adi"}),
the gini coefficient (\code{"gini"}), two versions of the neighborhood
deprivation index by Messer (\code{"ndi_m"}) and Powell and Wiley
(\code{"ndi_pw"}), and four versions of the social vulnerability
index (\code{"svi10"}, \code{"svi14"}, \code{"svi20"}, and \code{"svi20s"}).}
}
\value{
A tibble containing the raw 2022 American Community Sruvey data for
    the given index. Each tibble will contain observations for the 115
    counties in Missouri.
}
\description{
Create sample data for testing the package functionality.
}
\examples{
## load sample data
dep_sample_data(index = "ndi_m")

}
