% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_get_index.R
\name{dep_get_index}
\alias{dep_get_index}
\title{Calculate Deprivation Measures}
\usage{
dep_get_index(geography, variables, keep_subscales = FALSE,
    keep_components = FALSE, output, year, survey, state = NULL,
    county = NULL, territory = c("AS", "GU", "MP", "PR", "VI"),
    zcta = NULL, zcta3_method = NULL, geometry = FALSE,
    keep_geo_vars = FALSE, shift_geo = FALSE, units, debug = NULL)
}
\description{
Calculates various measures of deprivation.
}
