% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deprivation.R
\name{get_deprivation}
\alias{get_deprivation}
\title{Calculate Deprivation Measures}
\usage{
get_deprivation(geography, variables, keep_subscales = FALSE,
    keep_components = FALSE, output, year, survey, state = NULL,
    county = NULL, geometry = FALSE, keep_geo_vars = FALSE, shift_geo = FALSE,
    units, debug = NULL)
}
\arguments{
\item{geography}{A character scalar; one of \code{"state"}, \code{"county"}, or
\code{"tract"}}

\item{variables}{A character scalar or vector listing deprivation measures
to return. These include the gini coefficient (\code{gini})...}

\item{keep_subscales}{A logical scalar; if \code{FALSE} (defualt), subscales
for SVI, SVIp, and CCVI will not be returned. If \code{TRUE}, these components
will be returned.}

\item{keep_components}{A logical salar; if \code{FALSE} (defualt), only the
main output measures will be returned. If \code{TRUE}, all individual
Census variables for each requested measure of deprivation will be
returned. Be aware that this can create very large output tables.}

\item{output}{A character scalar; one of \code{"tidy"} (long output) or
\code{"wide"} depending on the type of data format you want.}

\item{year}{A numeric between 2010 and 2020}

\item{state}{A state name or FIPS code (may be either character or numeric
as well as scalar or vector)}

\item{county}{A county FIPS code or vector of FIPS codes}

\item{geometry}{A logical salar; if \code{FALSE} (defualt), tabular data are
returned. If \code{TRUE}, a \code{sf} class object will be returned.
This will only work when \code{output = "wide"}.}

\item{keep_geo_vars}{A logical salar; if \code{FALSE} (defualt), a minimum
number of spatial variables are returned. If \code{TRUE}, a full set of
TIGER/Line variables are returned. This is only relevant if
\code{geometry = TRUE}.}

\item{shift_geo}{A logical scalar; if \code{FALSE} (defualt), Alaska and
Hawaii will appear in their cartographically correct locations. If
\code{TRUE}, Alaska and Hawaii will be shifted south to facilitate
more straightforward national mapping.}

\item{units}{A character scalar; either \code{"mi"} (default; miles) or
\code{"km"} (kilometers)}

\item{debug}{Additional parameters for debugging and testing
\code{deprivateR}. The two styles are \code{"messages"} (will print
all messages from \code{tidycensus} and \code{sociome}) and
\code{"test"} (will use built-in data instead of making live API
calls).}

\item{survey}{A character scalar representing the Census product. It can
be any American Community Survey product (either \code{"acs1"},
\code{"acs3"}, or \code{"acs5"}). Note that \code{"acs3"} was
discontinued after 2013.}
}
\description{
Calculates various measures of deprivation.
}
